% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar_penal.R
\name{spinar_penal}
\alias{spinar_penal}
\title{Penalized (semiparametric) estimation of INAR(p) model.}
\usage{
spinar_penal(x, p, penal1 = 0, penal2 = 0)
}
\arguments{
\item{x}{[\code{integer}]\cr
vector of integer values corresponding to the data.}

\item{p}{[\code{integer(1)}]\cr
lag of the INAR(p) where \code{p in \{1,2\}}.}

\item{penal1}{default value penal1=0 meaning no penalization.}

\item{penal2}{default value penal2=0 meaning no penalization.}
}
\value{
parameters
}
\description{

}
\examples{
### data generated
dat <- spinar_sim(n = 100, p=1, alpha = 0.5, pmf = dpois(0:20,1))
## penalized spinar
spinar_penal(dat,1,0.5,0.5)
}
