% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar_penal_validation.R
\name{spinar_penal_wrapper}
\alias{spinar_penal_wrapper}
\title{Validation Penalized (semiparametric) estimation of INAR(p) model.}
\usage{
spinar_penal_wrapper(
  x,
  p,
  validation,
  penal1,
  penal2,
  over,
  folds = 10,
  init1 = 1,
  init2 = 1
)
}
\arguments{
\item{x}{vector of integer values corresponding to the data used for the `spinar_penal`.}

\item{p}{lag of the INAR(p) where \code{p in \{1,2\}}.}

\item{validation}{TRUE or FALSE depending on whether validation is wanted.}

\item{penal1}{penalization parameter for L1 penalization. It will be ignored if validation = TRUE and over = both, for validation = TRUE and over = L1 only one will be ignored; are mandatory if validation = FALSE.}

\item{penal2}{penalization parameter for L2 penalization. It will be ignored if validation = TRUE and over = both, for validation = TRUE and over = L2 only one will be ignored; are mandatory if validation = FALSE.}

\item{over}{mandatory if validation = TRUE, otherwise it will be ignored, answers whether validation for penal1 (L1) or penal2 (L2) or both (both) is wanted (error if over is not one of these three).}

\item{folds}{number of folds for cross validation.}

\item{init1}{initial values for penal1 in validation. Default value is init1 = 1.}

\item{init2}{initial values for penal2 in validation. Default value is init2 = 1.}
}
\value{
parameters
}
\description{

}
