% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar_sim.R
\name{spinar_sim}
\alias{spinar_sim}
\title{Simulation of (semi)parametric integer autoregressive models}
\usage{
spinar_sim(n, p, alpha, pmf, prerun = 500)
}
\arguments{
\item{n}{[\code{integer}(1)]\cr
total of values generated for the INAR(p).}

\item{p}{[\code{integer}(1)]\cr
lag of the INAR(p) where \code{p in \{1,2\}}.}

\item{alpha}{[\code{integer(p)}]\cr
alpha_1,...,alpha_p}

\item{pmf}{[\code{numeric}]\cr
probability mass function pmf0,..., pmfk where pmfi represent the pmf for the value i.}

\item{prerun}{[\code{integer(1)}]\cr
default parameter equal 500, the function `spinar_sim` uses `n + prerun` observations and then cut the first prerun observations.}
}
\value{
sim  [\code{integer(n)}]\cr
vector with n simulations for the INAR(p) generated.
}
\description{
Function to generate INAR(p), where p in {1,2} and we allow for general pmf's which can be generated parametrically or "manually" (semiparametric)
}
\examples{
spinar_sim(1000, 1, 0.5, dpois(0:20,1), prerun = 500)
spinar_sim(1000, 1, 0.5, c(0.1,0.5,0.4), prerun = 500)
spinar_sim(1000, 1, 0.5, c(0.1,0.5,0.2), prerun = 500)
spinar_sim(1000, 2, c(0.2, 0.3), dpois(0:20,1), prerun = 500)
spinar_sim(1000, 2, c(0.2, 0.3), c(0.1,0.5,0.4), prerun = 500)
spinar_sim(1000, 2, c(0.2, 0.3), c(0.1,0.5,0.2), prerun = 500)
}
