% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spINAR-package.R
\docType{package}
\name{spINAR-package}
\alias{spINAR}
\alias{spINAR-package}
\title{Estimation of semiparametric integer autoregressive models}
\description{
A set of functions used for the following tasks.
1. Simulate INAR(p) data with arbitrary innovation distributions (parametric family or probability mass function).
2. Semiparametric estimation of INAR(p) model.
3. Semiparametric INAR Boostrap.
4. Fully parametric estimation of INAR(p) model.
5. Penalized (semiparametric) estimation of INAR(p) model.
}
\references{
REFERENCES? arXiv paper?
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MFaymon/spINAR}
  \item Report bugs at \url{https://github.com/MFaymon/spINAR/issues}
}

}
\author{
\strong{Maintainer}: Maxime Faymonville \email{faymonville@statistik.tu-dortmund.de}

Authors:
\itemize{
  \item Javiera Riffo \email{javiera.riffo@tu-dortmund.de}
  \item Jonas Rieger \email{rieger@statistik.tu-dortmund.de} (\href{https://orcid.org/0000-0002-0007-4478}{ORCID})
}

}
