% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar.R
\name{spinar}
\alias{spinar}
\title{Parameter estimation of (semi)parametric integer autoregressive models}
\usage{
spinar(x, p)
}
\arguments{
\item{x}{[\code{integer}]\cr
vector of integer values corresponding to the data used for estimating the parameters via maximum likelihood.}

\item{p}{[\code{integer(1)}]\cr
lag of the INAR(p) where \code{p in \{1,2\}}}
}
\value{
parameters [\code{numeric(p+max(x)+1)}]\cr

Return the estimated parameters \code{(alpha_1, ..., alpha_p, pmf0, pmf1, ...)},
where \code{(alpha_1, ..., alpha_p)} are the estimated autoregressive coefficients
and \code{(pmf0, pmf1, ...)} are the estimated entries of the probability mass function of the innovation distribution, here \code{pmfi} denotes the entry of the integer \code{i} in \code{x}.
}
\description{
Performs a semiparametric estimation of the autoregressive parameters and the innovation distribution for the integer-valued autoregressive model of order \code{p} (INAR(\code{p})) for \code{p in {1,2}}, the estimation is conducted by maximizing the conditional likelihood of the model.
}
\examples{
## Parameter estimation for p = 1 and x as a vector of values.
spinar(c(2, 3, 1, 1, 1, 1, 1, 1, 3), 1)
### returns (alpha_1, pmf0, pmf1, pmf2, pmf3)
## Parameter estimation for p = 2 and data generated by rpois.
spinar(rpois(100, 0.8), 2)
### returns (alpha_1, alpha_2, pmf0, ...)
}
