% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar.R
\name{spinar}
\alias{spinar}
\title{Spinar(p)}
\usage{
spinar(dat, p)
}
\arguments{
\item{dat}{dataset}

\item{p}{lag}
}
\value{
Return the estimated parameters of the INAR(p). The first p-values
corresponds to (alpha_1,..., alpha_p)
the next values corresponds to the probability mass function (pmf0, pmf1, ...)
}
\description{
This function takes a data set and the lag of INAR process
}
\examples{
spinar(c(2,3,1,1,1,1,1,1,3), 1)
2.362001e-11 4.205347e-09 6.128583e-01 5.021233e-10 3.871417e-01

}
