% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar.R
\name{spinar}
\alias{spinar}
\title{Parameter estimation of (semi)parametric integer autoregressive models}
\usage{
spinar(x, p)
}
\arguments{
\item{x}{[\code{integer}]\cr
EXPLAIN IN DETAIL!.}

\item{p}{[\code{integer(1)}]\cr
Lag. EXPLAIN IN DETAIL!}
}
\value{
[\code{numeric(p+??)}]\cr
Estimated parameters \code{(alpha_1,..., alpha_p, pmf0, pmf1, ...)}. EXPLAIN IN DETAIL!
}
\description{
Performs bla bla.... EXPLAIN IN DETAIL!
}
\examples{
# EXPLAIN IN DETAIL!
spinar(c(2,3,1,1,1,1,1,1,3), 1)
# 2nd EXAMPLE!
}
