% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar_sp_boot.R
\name{spinar_boot}
\alias{spinar_boot}
\title{Semiparametric Integer Autoregressive Boostrap model.}
\usage{
spinar_boot(x, p, B)
}
\arguments{
\item{x}{[\code{integer}]\cr
vector of integer values corresponding to the data used for estimate the parameter using `spinar` function.}

\item{p}{[\code{integer(1)}]\cr
lag of the INAR(p) where \code{p in \{1,2\}}}

\item{B}{[\code{integer(1)}]\cr}
}
\value{
parameters
}
\description{

}
\examples{
### simulating data of INAR(1)
dat <- spinar_sim(n=1000, p = 1, alpha = 0.3, pmf = dpois(0:10,1.5))
## Spinar boosting
spinar_boot(dat, 1, 10)
### simulating data of INAR(2)
dat <- spinar_sim(n=500, p = 2, alpha = c(0.3, 0.2), pmf = dpois(0:10,1.5))
spinar_boot(dat, 2, 10)
}
