% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar_sp_boot.R, R/spinar_boot.R
\name{spinar_boot}
\alias{spinar_boot}
\title{Semiparametric Integer Autoregressive Bootstrap Procedure}
\usage{
spinar_boot(x, p, B, setting, type = NA, distr = NA, M = 100)

spinar_boot(x, p, B, setting, type = NA, distr = NA, M = 100)
}
\arguments{
\item{x}{[\code{integer}]\cr
vector of integer values corresponding to the data}

\item{p}{[\code{integer(1)}]\cr
order of the INAR model, where \code{p in \{1,2\}}}

\item{B}{[\code{integer(1)}]\cr
number of bootstrap repetitions}

\item{setting}{[\code{string(1)}]\cr
setting of estimation in \code{\{"sp", "p"\}}
'sp': semiparametric setting
'p' : parametric setting}

\item{type}{[\code{string(1)}]\cr
type of estimation in \code{\{"mom", "ml"\}}
'mom': moment estimation
'ml' : maximum likelihood estimation}

\item{distr}{[\code{string(1)}]\cr
parametric family of distribution in  \code{\{'poi', 'geo', 'nb'\}}
\code{'poi'}: Poisson distribution with parameter \code{lambda}
\code{'geo'}: Geometric distribution with parameter \code{prob}
\code{'nb'}: Negative binomial distribution with parameters \code{r} and \code{prob}}

\item{M}{[\code{integer(1)}]\cr
upper limit for innovations}
}
\value{
bootstrap estimated parameters (INAR coefficients and pmf)

bootstrap estimated parameters (INAR coefficients and pmf (or estimated innovation parameters))
}
\description{



}
\examples{
### simulating data of INAR(1)
# dat <- spinar_sim(n=1000, p = 1, alpha = 0.3, pmf = dpois(0:10,1.5))
### semiparametric INAR bootstrap
# spinar_boot(dat, 1, 100)
### simulating data of INAR(2)
# dat <- spinar_sim(n=500, p = 2, alpha = c(0.3, 0.2), pmf = dpois(0:10,1.5))
### semiparametric INAR bootstrap
# spinar_boot(dat, 2, 100)
### simulating data of INAR(1)
# dat <- spinar_sim(n=1000, p = 1, alpha = 0.3, pmf = dpois(0:10,1.5))
### semiparametric INAR bootstrap
# spinar_boot(dat, 1, 100)
### simulating data of INAR(2)
# dat <- spinar_sim(n=500, p = 2, alpha = c(0.3, 0.2), pmf = dpois(0:10,1.5))
### semiparametric INAR bootstrap
# spinar_boot(x=dat, p=2, B=100, setting="sp")
### parametric Poisson INAR bootstrap
spinar_boot(x=dat, p=2, B=100, setting="p", type="mom", distr="poi")

}
