% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar_est.R
\name{spinar_est}
\alias{spinar_est}
\title{Semiparametric Estimation of INAR Models}
\usage{
spinar_est(x, p)
}
\arguments{
\item{x}{[\code{integer}]\cr
vector of integer values corresponding to the data}

\item{p}{[\code{integer(1)}]\cr
order of the INAR model, where \code{p in \{1,2\}}}
}
\value{
estimated parameters \code{(alpha_1, ..., alpha_p, pmf[0], pmf[1], ...)},
where \code{(alpha_1, ..., alpha_p)} are the estimated autoregressive coefficients
and \code{(pmf[0], pmf[1], ...)} are the estimated entries of the probability mass function of the innovation distribution,
where \code{pmf[i]} denotes the probability of observing value i
}
\description{
Performs a semiparametric estimation of the autoregressive parameters and the innovation distribution for INAR models of order \code{p},
 \code{p in {1,2}}, the estimation is conducted by maximizing the conditional likelihood of the model.
}
\examples{
## Parameter estimation for p = 1 and x as a vector of values.
dat <- spinar_sim(100, 1, 0.5, dpois(0:10,1))
spinar_est(dat, 1)
### returns (alpha_1, pmf[0], pmf[1],...)
## Parameter estimation for p = 2 and data generated from a poisson distribution with parameter 0.8.
dat <- spinar_sim(100, 2, 0.2, 0.3, dpois(0:10,1))
spinar_est(dat, 2)
### returns (alpha_1, alpha_2, pmf[0], pmf[1], ...)
}
