% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar_est_param.R
\name{spinar_est_param}
\alias{spinar_est_param}
\title{Parametric estimation of INAR models}
\usage{
spinar_est_param(x, p, type, distr)
}
\arguments{
\item{x}{[\code{integer}]\cr
vector of integer values corresponding to the data}

\item{p}{[\code{integer(1)}]\cr
order of the INAR model, where \code{p in \{1,2\}}}

\item{distr}{}
}
\value{

}
\description{
for now only moment-based
}
